+++ packages/maestro/lib/src/maestro.dart
@@ -8,7 +8,7 @@ import 'package:maestro/src/inspector.dart';
 /// Its descendants can read the value or write a new value. When the value is
 /// replaced, all descendants who listen a specific [Maestro] will be rebuilt.
 class Maestro<T> extends StatefulWidget implements Relocatable<Maestro<T>> {
-  /// Stores the [initialValue] and exposes it to its descendants.
+  /// Stores the [value] and exposes it to its descendants.
   ///
   /// `equalityComparer` can optionally be passed to avoid unnecessarily
   /// rebuilding dependents when [Maestro] is rebuilt but its `value`
@@ -16,14 +16,14 @@ class Maestro<T> extends StatefulWidget implements Relocatable<Maestro<T>> {
   ///
   /// Defaults to `(previous, next) => previous != next`.
   const Maestro(
-    this.initialValue, {
+    this.value, {
     Key key,
     this.equalityComparer,
     this.child,
   }) : super(key: key);
 
   /// The initial value held by this widget.
-  final T initialValue;
+  final T value;
 
   /// Used to compare old and new values in order to rebuild its descendants
   /// only when these values are not considered equals.
@@ -39,7 +39,7 @@ class Maestro<T> extends StatefulWidget implements Relocatable<Maestro<T>> {
   @override
   Maestro<T> copyWithNewChild(Widget newChild) {
     return Maestro<T>(
-      initialValue,
+      value,
       key: key,
       equalityComparer: equalityComparer,
       child: newChild,
@@ -125,8 +125,6 @@ class Maestro<T> extends StatefulWidget implements Relocatable<Maestro<T>> {
 }
 
 class _MaestroState<T> extends State<Maestro<T>> implements Score {
-  MaestroInspector _inspector;
-
   T get value => _value;
   T _value;
   void _dispatch(_Wrapper<T> wrapper) {
@@ -135,17 +133,28 @@ class _MaestroState<T> extends State<Maestro<T>> implements Score {
     setState(() {
       _value = wrapper.value;
     });
-    _inspector?.onAction(oldValue, value, wrapper.action);
+    _inspect(oldValue, value, wrapper.action);
+  }
+
+  void _inspect(T oldValue, T value, Object action) {
+    bool bubbling = true;
+    _MaestroScope<MaestroInspector> inspector =
+        context._getMaestroScope<MaestroInspector>();
+    while (inspector != null && bubbling) {
+      bubbling = !inspector.value.onAction(oldValue, value, action);
+      inspector = inspector.state.context._getMaestroScope<MaestroInspector>();
+    }
   }
 
   @override
   void initState() {
     super.initState();
-    _value = widget.initialValue;
+    _value = widget.value;
     final T value = _value;
     if (value is Performer) {
+      value.attach(this);
       WidgetsBinding.instance.addPostFrameCallback((_) {
-        value.attach(this);
+        value.play();
       });
     }
   }
@@ -153,17 +162,18 @@ class _MaestroState<T> extends State<Maestro<T>> implements Score {
   @override
   void didUpdateWidget(Maestro<T> oldWidget) {
     super.didUpdateWidget(oldWidget);
-    final T newValue = widget.initialValue;
-    if (newValue is! Performer && _updateShouldNotify(value, newValue)) {
-      // Only the widgets where values are not a Performer update their value.
+    final T newValue = widget.value;
+    if (_updateShouldNotify(value, newValue)) {
+      if (newValue is Performer) {
+        newValue.attach(this);
+        final T oldValue = value;
+        if (oldValue is Performer) {
+          newValue.remix(oldValue);
+          oldValue.detach();
+        }
+      }
       _dispatch(_Wrapper<T>(newValue, const WidgetUpdatedAction()));
     }
-  }
-
-  @override
-  void didChangeDependencies() {
-    super.didChangeDependencies();
-    _inspector = Maestro.listen<MaestroInspector>(context);
   }
 
   @override