+++ packages/maestro/test/composer_test.dart
@@ -4,15 +4,6 @@ import 'package:maestro/maestro.dart';
 
 import 'common.dart';
 
-class _DelegatedComposer with Composer {
-  _DelegatedComposer(this._play);
-
-  final VoidCallback _play;
-
-  @override
-  void play() => _play();
-}
-
 class _FetchWhenAttachedComposer<T> with Composer {
   _FetchWhenAttachedComposer(this.fetch);
   final ValueChanged<T> fetch;
@@ -21,50 +12,15 @@ class _FetchWhenAttachedComposer<T> with Composer {
   void play() {
     fetch(read<T>());
   }
+
+  @override
+  void remix(_FetchWhenAttachedComposer old) {
+    fetch(read<T>());
+  }
 }
 
 void main() {
   group('Composer', () {
-    testWidgets('attached when added in the tree', (tester) async {
-      final List<bool> play = [false, false];
-
-      final List<_DelegatedComposer> composers = List.generate(
-        2,
-        (i) => _DelegatedComposer(
-          () {
-            play[i] = true;
-          },
-        ),
-      );
-
-      await tester.pumpWidget(
-        Maestros(
-          [
-            const Maestro(1),
-            Maestro(composers[0], key: UniqueKey()),
-          ],
-          child: const SizedBox(),
-        ),
-      );
-
-      expect(play[0], equals(true));
-      expect(play[1], equals(false));
-
-      await tester.pumpWidget(
-        Maestros(
-          [
-            const Maestro(1),
-            Maestro(composers[1], key: UniqueKey()),
-          ],
-          child: const SizedBox(),
-        ),
-      );
-
-      // The new composer is attached the Maestro is added with a new key.
-      expect(play[0], equals(true));
-      expect(play[1], equals(true));
-    });
-
     testWidgets('can read an ancestor maestro when attached', (tester) async {
       int value;
 
@@ -83,11 +39,8 @@ void main() {
       await tester.pumpWidget(
         Maestros(
           [
-            Maestro(2, key: UniqueKey()),
-            Maestro(
-              _FetchWhenAttachedComposer<int>((int x) => value = x),
-              key: UniqueKey(),
-            ),
+            const Maestro(2),
+            Maestro(_FetchWhenAttachedComposer<int>((int x) => value = x)),
           ],
           child: const SizedBox(),
         ),