+++ packages/maestro/test/inspector_test.dart
@@ -10,8 +10,9 @@ void main() {
         (tester) async {
       final List<String> logs = <String>[];
 
-      void onAction<T>(T old, T value, Object action) {
+      bool onAction<T>(T old, T value, Object action) {
         logs.add('$old => $value');
+        return true;
       }
 
       final MaestroInspector inspector = MaestroInspector(onAction);
@@ -41,8 +42,9 @@ void main() {
     testWidgets('method is called with action name', (tester) async {
       final List<String> logs = <String>[];
 
-      void onAction<T>(T old, T value, Object action) {
+      bool onAction<T>(T old, T value, Object action) {
         logs.add('$action: $old => $value');
+        return true;
       }
 
       final MaestroInspector inspector = MaestroInspector(onAction);
@@ -67,6 +69,92 @@ void main() {
 
       composer.store('b', 'StringStorage');
       expect(logs, <String>['IntStorage: 1 => 2', 'StringStorage: a => b']);
+    });
+
+    testWidgets('methods are called in order', (tester) async {
+      final List<String> logs = <String>[];
+
+      bool onAction01<T>(T old, T value, Object action) {
+        logs.add('01 $action: $old => $value');
+        return false;
+      }
+
+      bool onAction02<T>(T old, T value, Object action) {
+        logs.add('02 $action: $old => $value');
+        return false;
+      }
+
+      final MaestroInspector inspector01 = MaestroInspector(onAction01);
+      final MaestroInspector inspector02 = MaestroInspector(onAction02);
+      final DefaultComposer composer = DefaultComposer();
+
+      await tester.pumpWidget(
+        Maestros(
+          [
+            Maestro(inspector01),
+            const Maestro(1),
+            Maestro(inspector02),
+            const Maestro('a'),
+            Maestro(composer),
+          ],
+          child: const SizedBox(),
+        ),
+      );
+
+      expect(logs, isEmpty);
+
+      composer.store(2, 'IntStorage');
+      expect(logs, <String>['01 IntStorage: 1 => 2']);
+
+      composer.store('b', 'StringStorage');
+      expect(logs, <String>[
+        '01 IntStorage: 1 => 2',
+        '02 StringStorage: a => b',
+        '01 StringStorage: a => b',
+      ]);
+    });
+
+    testWidgets('methods are called in order and cascading is respected',
+        (tester) async {
+      final List<String> logs = <String>[];
+
+      bool onAction01<T>(T old, T value, Object action) {
+        logs.add('01 $action: $old => $value');
+        return false;
+      }
+
+      bool onAction02<T>(T old, T value, Object action) {
+        logs.add('02 $action: $old => $value');
+        return true;
+      }
+
+      final MaestroInspector inspector01 = MaestroInspector(onAction01);
+      final MaestroInspector inspector02 = MaestroInspector(onAction02);
+      final DefaultComposer composer = DefaultComposer();
+
+      await tester.pumpWidget(
+        Maestros(
+          [
+            Maestro(inspector01),
+            const Maestro(1),
+            Maestro(inspector02),
+            const Maestro('a'),
+            Maestro(composer),
+          ],
+          child: const SizedBox(),
+        ),
+      );
+
+      expect(logs, isEmpty);
+
+      composer.store(2, 'IntStorage');
+      expect(logs, <String>['01 IntStorage: 1 => 2']);
+
+      composer.store('b', 'StringStorage');
+      expect(logs, <String>[
+        '01 IntStorage: 1 => 2',
+        '02 StringStorage: a => b',
+      ]);
     });
   });
 }